+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # SceneKit Neumorphism
 
+* 0.0.5
+
+    + Added `@EnvironmentObject` to collect geometry data
+    + Added `UIViewRepresentableContext.coordinator` to receive `SceneKit` scene updates  
+
 * 0.0.0 - 0.0.4
 
     + Prototype